+++ src/account/AccountAllowanceAdjustTransaction.js
@@ -1,6 +1,24 @@
-import Transaction, {
-    TRANSACTION_REGISTRY,
-} from "../transaction/Transaction.js";
+/*-
+ * ‌
+ * Hedera JavaScript SDK
+ * ​
+ * Copyright (C) 2020 - 2022 Hedera Hashgraph, LLC
+ * ​
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ‍
+ */
+
+import Transaction from "../transaction/Transaction.js";
 import AccountId from "./AccountId.js";
 import TokenId from "../token/TokenId.js";
 import NftId from "../token/NftId.js";
@@ -18,7 +36,6 @@ import * as util from "../util.js";
  * @typedef {import("@hashgraph/proto").proto.TransactionBody} HashgraphProto.proto.TransactionBody
  * @typedef {import("@hashgraph/proto").proto.ITransactionBody} HashgraphProto.proto.ITransactionBody
  * @typedef {import("@hashgraph/proto").proto.ITransactionResponse} HashgraphProto.proto.ITransactionResponse
- * @typedef {import("@hashgraph/proto").proto.ICryptoAdjustAllowanceTransactionBody} HashgraphProto.proto.ICryptoAdjustAllowanceTransactionBody
  * @typedef {import("@hashgraph/proto").proto.IAccountID} HashgraphProto.proto.IAccountID
  */
 
@@ -31,6 +48,7 @@ import * as util from "../util.js";
  */
 
 /**
+ * @deprecated - No longer supported via Hedera Protobufs
  * Change properties for the given account.
  */
 export default class AccountAllowanceAdjustTransaction extends Transaction {
@@ -63,51 +81,6 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
          */
         this._nftAllowances =
             props.nftAllowances != null ? props.nftAllowances : [];
-    }
-
-    /**
-     * @internal
-     * @param {HashgraphProto.proto.ITransaction[]} transactions
-     * @param {HashgraphProto.proto.ISignedTransaction[]} signedTransactions
-     * @param {TransactionId[]} transactionIds
-     * @param {AccountId[]} nodeIds
-     * @param {HashgraphProto.proto.ITransactionBody[]} bodies
-     * @returns {AccountAllowanceAdjustTransaction}
-     */
-    static _fromProtobuf(
-        transactions,
-        signedTransactions,
-        transactionIds,
-        nodeIds,
-        bodies
-    ) {
-        const body = bodies[0];
-        const allowanceAdjust =
-            /** @type {HashgraphProto.proto.ICryptoAdjustAllowanceTransactionBody} */ (
-                body.cryptoAdjustAllowance
-            );
-
-        return Transaction._fromProtobufTransactions(
-            new AccountAllowanceAdjustTransaction({
-                hbarAllowances: (allowanceAdjust.cryptoAllowances != null
-                    ? allowanceAdjust.cryptoAllowances
-                    : []
-                ).map((adjust) => HbarAllowance._fromProtobuf(adjust)),
-                tokenAllowances: (allowanceAdjust.tokenAllowances != null
-                    ? allowanceAdjust.tokenAllowances
-                    : []
-                ).map((adjust) => TokenAllowance._fromProtobuf(adjust)),
-                nftAllowances: (allowanceAdjust.nftAllowances != null
-                    ? allowanceAdjust.nftAllowances
-                    : []
-                ).map((adjust) => TokenNftAllowance._fromProtobuf(adjust)),
-            }),
-            transactions,
-            signedTransactions,
-            transactionIds,
-            nodeIds,
-            bodies
-        );
     }
 
     /**
@@ -159,6 +132,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
      * @param {number | string | Long | LongObject | BigNumber | Hbar} amount
@@ -174,6 +148,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
      * @param {number | string | Long | LongObject | BigNumber | Hbar} amount
@@ -196,6 +171,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {TokenId | string} tokenId
      * @param {AccountId | string} spenderAccountId
      * @param {Long | number} amount
@@ -245,6 +221,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {TokenId | string} tokenId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -261,6 +238,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {TokenId | string} tokenId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -336,6 +314,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {NftId | string} nftId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -354,6 +333,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {NftId | string} nftId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -386,6 +366,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {TokenId | string} tokenId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -405,6 +386,7 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {TokenId | string} tokenId
      * @param {AccountId | string} ownerAccountId
      * @param {AccountId | string} spenderAccountId
@@ -484,36 +466,32 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
      * @param {HashgraphProto.proto.ITransaction} request
      * @returns {Promise<HashgraphProto.proto.ITransactionResponse>}
      */
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     _execute(channel, request) {
-        return channel.crypto.adjustAllowances(request);
+        return Promise.reject(
+            new Error("This feature has been deprecated for this class.")
+        );
     }
 
+    // eslint-disable-next-line jsdoc/require-returns-check
     /**
+     * @deprecated
      * @override
      * @protected
      * @returns {NonNullable<HashgraphProto.proto.TransactionBody["data"]>}
      */
     _getTransactionDataCase() {
-        return "cryptoAdjustAllowance";
+        throw new Error("This feature has been deprecated for this class.");
     }
 
+    // eslint-disable-next-line jsdoc/require-returns-check
     /**
      * @override
      * @protected
-     * @returns {HashgraphProto.proto.ICryptoAdjustAllowanceTransactionBody}
+     * @returns {Response} Error
      */
     _makeTransactionData() {
-        return {
-            cryptoAllowances: this._hbarAllowances.map((adjust) =>
-                adjust._toProtobuf()
-            ),
-            tokenAllowances: this._tokenAllowances.map((adjust) =>
-                adjust._toProtobuf()
-            ),
-            nftAllowances: this._nftAllowances.map((adjust) =>
-                adjust._toProtobuf()
-            ),
-        };
+        throw new Error("This feature has been deprecated.");
     }
 
     /**
@@ -526,9 +504,3 @@ export default class AccountAllowanceAdjustTransaction extends Transaction {
         return `AccountAllowanceAdjustTransaction:${timestamp.toString()}`;
     }
 }
-
-TRANSACTION_REGISTRY.set(
-    "cryptoAdjustAllowance",
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    AccountAllowanceAdjustTransaction._fromProtobuf
-);