+++ src/account/TokenAllowance.js
@@ -1,3 +1,23 @@
+/*-
+ * ‌
+ * Hedera JavaScript SDK
+ * ​
+ * Copyright (C) 2020 - 2022 Hedera Hashgraph, LLC
+ * ​
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ‍
+ */
+
 import TokenId from "../token/TokenId.js";
 import AccountId from "./AccountId.js";
 import Long from "long";
@@ -5,7 +25,6 @@ import Long from "long";
 /**
  * @namespace proto
  * @typedef {import("@hashgraph/proto").proto.IGrantedTokenAllowance} HashgraphProto.proto.IGrantedTokenAllowance
- * @typedef {import("@hashgraph/proto").proto.ITokenRemoveAllowance} HashgraphProto.proto.ITokenRemoveAllowance
  * @typedef {import("@hashgraph/proto").proto.ITokenAllowance} HashgraphProto.proto.ITokenAllowance
  * @typedef {import("@hashgraph/proto").proto.ITokenID} HashgraphProto.proto.ITokenID
  * @typedef {import("@hashgraph/proto").proto.IAccountID} HashgraphProto.proto.IAccountID
@@ -109,29 +128,6 @@ export default class TokenAllowance {
                 allowance.amount != null
                     ? Long.fromValue(/** @type {Long} */ (allowance.amount))
                     : null,
-        });
-    }
-
-    /**
-     * @internal
-     * @param {HashgraphProto.proto.ITokenRemoveAllowance} allowance
-     * @returns {TokenAllowance}
-     */
-    static _fromRemoveAllowance(allowance) {
-        return new TokenAllowance({
-            tokenId: TokenId._fromProtobuf(
-                /** @type {HashgraphProto.proto.ITokenID} */ (allowance.tokenId)
-            ),
-            spenderAccountId: null,
-            ownerAccountId:
-                allowance.owner != null
-                    ? AccountId._fromProtobuf(
-                          /**@type {HashgraphProto.proto.IAccountID}*/ (
-                              allowance.owner
-                          )
-                      )
-                    : null,
-            amount: null,
         });
     }